+++ layabauth/flask.py
@@ -1,24 +1,34 @@
 import logging
 import functools
+import json
 
 import flask
 import werkzeug
-import jwt
-import oauth2helper
+from jose import exceptions, jws
 
-from layabauth._http import _get_token
+from layabauth._http import _get_token, validate, keys
 
 
-def requires_authentication(identity_provider_url: str):
+def requires_authentication(jwks_uri: str):
+    """
+    Ensure that a valid JWT is received before entering the annotated endpoint.
+
+    :param jwks_uri: The JWKs URI as defined in .well-known.
+    For more information on JWK, refer to https://tools.ietf.org/html/rfc7517
+        * Azure Active Directory: https://sts.windows.net/common/discovery/keys
+        * Microsoft Identity Platform: https://sts.windows.net/common/discovery/keys
+    """
+
     def decorator(func):
         @functools.wraps(func)
         def wrapper(*func_args, **func_kwargs):
             try:
                 flask.g.token = _get_token(flask.request.headers)
-                json_header, flask.g.token_body = oauth2helper.validate(
-                    flask.g.token, identity_provider_url
-                )
-            except jwt.PyJWTError as e:
+                if not flask.g.token:
+                    raise werkzeug.exceptions.Unauthorized()
+                key = keys(jwks_uri)
+                flask.g.token_body = validate(flask.g.token, key)
+            except exceptions.JOSEError as e:
                 raise werkzeug.exceptions.Unauthorized(description=str(e)) from e
             return func(*func_args, **func_kwargs)
 
@@ -30,10 +40,14 @@ def requires_authentication(identity_provider_url: str):
 def _extract_token_body() -> dict:
     if getattr(flask.g, "token_body", None):
         return flask.g.token_body
+
+    token = _get_token(flask.request.headers)
+    if not token:
+        return {}
+
     try:
-        json_header, json_body = oauth2helper.decode(_get_token(flask.request.headers))
-        return json_body
-    except jwt.PyJWTError:
+        return json.loads(jws.get_unverified_claims(token=token))
+    except exceptions.JOSEError:
         return {}