+++ layabauth/starlette.py
@@ -1,7 +1,5 @@
 from typing import Optional, Tuple
 
-import jwt
-import oauth2helper
 from starlette.authentication import (
     AuthenticationBackend,
     AuthCredentials,
@@ -9,8 +7,9 @@ from starlette.authentication import (
     AuthenticationError,
 )
 from starlette.requests import Request
+from jose import exceptions
 
-from layabauth._http import _get_token
+from layabauth._http import _get_token, validate, keys
 
 
 class OAuth2IdTokenBackend(AuthenticationBackend):
@@ -18,16 +17,16 @@ class OAuth2IdTokenBackend(AuthenticationBackend):
     Handle authentication via OAuth2 id-token (implicit flow, authorization code, with or without PKCE)
     """
 
-    def __init__(
-        self, identity_provider_url: str, create_user: callable, scopes: callable
-    ):
+    def __init__(self, jwks_uri: str, create_user: callable, scopes: callable):
         """
-        :param identity_provider_url: URL to retrieve the keys.
+        :param jwks_uri: The JWKs URI as defined in .well-known.
+        For more information on JWK, refer to https://tools.ietf.org/html/rfc7517
             * Azure Active Directory: https://sts.windows.net/common/discovery/keys
+            * Microsoft Identity Platform: https://sts.windows.net/common/discovery/keys
         :param create_user: callable receiving the token and the decoded token body and returning a starlette.BaseUser instance.
         :param scopes: callable receiving the token and the decoded token body and returning the list of associated scopes str.
         """
-        self.identity_provider_url = identity_provider_url
+        self.jwks_uri = jwks_uri
         self.create_user = create_user
         self.scopes = scopes
 
@@ -39,10 +38,9 @@ class OAuth2IdTokenBackend(AuthenticationBackend):
             return  # Consider that user is not authenticated
 
         try:
-            json_header, json_body = oauth2helper.validate(
-                token, self.identity_provider_url
-            )
-        except jwt.PyJWTError as e:
+            key = keys(self.jwks_uri)
+            json_body = validate(token, key)
+        except exceptions.JOSEError as e:
             raise AuthenticationError(str(e)) from e
 
         return (