+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.1] - 2020-04-28
+### Fixed
+- Drop oauth2helper in favor of python-jose to handle all kind of tokens.
+
+### Changed
+- `identity_provider_url` has been renamed to `jwks_uri` to match the key in .well-known
+- `jwks_uri` pytest fixture needs to be provided in addition to `token_body`.
+- Flask tests will require a fake token to be provided in headers (unless you want to test behavior without providing a token).
+
 ## [4.0.0] - 2020-04-20
 ### Changed
 - Flask specifics are now within layabauth.flask.
@@ -18,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/Colin-b/layabauth/compare/v4.0.0...HEAD
+[Unreleased]: https://github.com/Colin-b/layabauth/compare/v4.0.1...HEAD
+[4.0.1]: https://github.com/Colin-b/layabauth/compare/v4.0.0...v4.0.1
 [4.0.0]: https://github.com/Colin-b/layabauth/compare/v3.2.0...v4.0.0
 [3.2.0]: https://github.com/Colin-b/layabauth/releases/tag/v3.2.0